# Fix s3 upload performance regression