--- a/objstore_test.go
+++ b/objstore_test.go
@@ -210,9 +210,7 @@ func TestDownloadUploadDirConcurrency(t *testing.T) {
 func TestTimingReader(t *testing.T) {
 	m := WrapWithMetrics(NewInMemBucket(), nil, "")
 	r := bytes.NewReader([]byte("hello world"))
-
-	tr := NopCloserWithSize(r)
-	tr = newTimingReader(tr, "", m.opsDuration, m.opsFailures, func(err error) bool {
+	tr := newTimingReader(r, "", m.opsDuration, m.opsFailures, func(err error) bool {
 		return false
 	}, m.opsFetchedBytes, m.opsTransferredBytes)
 
@@ -231,6 +229,13 @@ func TestTimingReader(t *testing.T) {
 
 	testutil.Ok(t, err)
 	testutil.Equals(t, int64(11), size)
+
+	// Given the reader was bytes.Reader it should both implement io.Seeker and io.ReaderAt.
+	_, isSeeker := tr.(io.Seeker)
+	testutil.Assert(t, isSeeker)
+
+	_, isReaderAt := tr.(io.ReaderAt)
+	testutil.Assert(t, isReaderAt)
 }
 
 func TestTimingReader_ShouldCorrectlyWrapFile(t *testing.T) {
@@ -324,35 +329,6 @@ func (b *uploadTrackerTestBucket) Upload(ctx context.Context, name string, r io.
 	return nil
 }
 
-func TestTimingReader_ShouldWrapSeeker(t *testing.T) {
-	m := WrapWithMetrics(NewInMemBucket(), nil, "")
-	r := bytes.NewReader([]byte("hello world"))
-
-	tr := nopSeekerCloserWithSize(r).(io.ReadCloser)
-	tr = newTimingReader(tr, "", m.opsDuration, m.opsFailures, func(err error) bool {
-		return false
-	}, m.opsFetchedBytes, m.opsTransferredBytes)
-
-	size, err := TryToGetSize(tr)
-
-	testutil.Ok(t, err)
-	testutil.Equals(t, int64(11), size)
-
-	smallBuf := make([]byte, 4)
-	n, err := io.ReadFull(tr, smallBuf)
-	testutil.Ok(t, err)
-	testutil.Equals(t, 4, n)
-
-	// Verify that size is still the same, after reading 4 bytes.
-	size, err = TryToGetSize(tr)
-
-	testutil.Ok(t, err)
-	testutil.Equals(t, int64(11), size)
-
-	_, ok := tr.(io.Seeker)
-	testutil.Equals(t, true, ok)
-}
-
 func TestDownloadDir_CleanUp(t *testing.T) {
 	b := unreliableBucket{
 		Bucket:  NewInMemBucket(),
@@ -384,17 +360,4 @@ func (b unreliableBucket) Get(ctx context.Context, name string) (io.ReadCloser, 
 		return nil, errors.Errorf("some error message")
 	}
 	return b.Bucket.Get(ctx, name)
-}
-
-type nopSeekerCloserWithObjectSize struct{ io.Reader }
-
-func (nopSeekerCloserWithObjectSize) Close() error                 { return nil }
-func (n nopSeekerCloserWithObjectSize) ObjectSize() (int64, error) { return TryToGetSize(n.Reader) }
-
-func (n nopSeekerCloserWithObjectSize) Seek(offset int64, whence int) (int64, error) {
-	return n.Reader.(io.Seeker).Seek(offset, whence)
-}
-
-func nopSeekerCloserWithSize(r io.Reader) io.ReadSeekCloser {
-	return nopSeekerCloserWithObjectSize{r}
 }